track and trace ui

+++ src/App.jsx
@@ -9,6 +9,7 @@ import AccountSettings from "./pages/Admin/AccountSettings.jsx";
 import AddProduct from "./pages/Product/AddProduct.jsx";
 import { useEffect } from "react";
 import { useLocation } from "react-router-dom";
+import TrackAndTrace from "./pages/TrackAndTrace/TrackAndTrace.jsx";
 
 const roles = {
   Admin: "Admin",
@@ -20,7 +21,10 @@ function App() {
 
   useEffect(() => {
     const clearLocalStorage = () => {
-      if (location.pathname !== "/admin/addProduct") {
+      if (
+        location.pathname !== "/admin/addProduct" &&
+        location.pathname !== "/track-and-trace"
+      ) {
         localStorage.clear();
       }
     };
@@ -41,6 +45,7 @@ function App() {
       <Route path="/unauthorized" element={<Unauthorized />} />
       <Route path="/01/:gtin/10/:lot" element={<Home />} />
       <Route path="/admin" element={<Dashboard />} />
+      <Route path="/track-and-trace" element={<TrackAndTrace />} />
 
       <Route element={<RequireAuth allowedRoles={[roles.Admin]} />}>
         <Route path="/admin/settings" element={<AccountSettings />} />
